refactor(cli): move _restart.py into cli/update/ (issue #4670 Part B)
The 12-line process-restart helper has only two consumers, both
inside cli/update/ (_update_checks.py, _update.py). Moving it
into the package keeps the restarter next to its callers and
leaves it with its existing basename (cli/update/'s naming is
mixed: _obligation_repair.py/_transaction.py unprefixed,
_update*.py prefixed).

- Rename cli/_restart.py -> cli/update/_restart.py
- Convert _update_checks.py and _update.py imports to absolute
- Update tests/cli/test_restart.py docstring and all import paths
